# Conflicts:
#	.idea/EetakemonConsola.iml
#	WEB/forms/MenuPrincipal.html
#	WEB/forms/Registrarse.html
#	WEB/index.html
#	src/main/java/Controlador/EetakemonService.java
#	src/main/java/Controlador/UserService.java
#	src/main/java/Modelo/Eetakemon/EetakemonManager.java
#	src/main/java/Modelo/Security/TrippleDes.java
#	src/main/java/Modelo/User/UserManager.java
#	src/test/java/Test.java

+++ src/main/java/Modelo/User/UserManager.java
@@ -1,11 +1,16 @@
 package Modelo.User;
 
+import Modelo.Exceptions.NotSuchPrivilegeException;
+import Modelo.Exceptions.UnauthorizedException;
+import Modelo.Security.AuthenticationManager;
 import Modelo.Security.TrippleDes;
+import Modelo.Security.Verification;
 import org.apache.log4j.Logger;
 
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
+import javax.ws.rs.core.HttpHeaders;
 import java.util.List;
 import java.util.Properties;
 
@@ -13,6 +18,7 @@ import java.util.Properties;
 public class UserManager {
     private TrippleDes td;
     private final static Logger logger = Logger.getLogger(UserManager.class);//
+    private AuthenticationManager authManager;
 
     public UserManager(){
         try{
@@ -21,6 +27,7 @@ public class UserManager {
         catch (Exception e){
 
         }
+        authManager=new AuthenticationManager();
     }
 
     /*
@@ -47,12 +54,12 @@ public class UserManager {
     /*
     Funcion registrarse
     code 0 = registrado
-    code 1 = eusuario ya utilizado
-    code 2 = error al registrarse
+    code 1 = usuario ya utilizado
+    code 2 = bad request
      */
     public int register(User user){
         boolean a=false,b=true;
-        int code=2;
+        int code;
         a=checkNullFields(user);
         if(!a) {
             b = user.checkUserExistent(user.getEmail());
@@ -67,21 +74,37 @@ public class UserManager {
             else{
                 code=1;
             }
-        }
+        }else{code=2;}
 
         return code;
     }
 
-    public boolean updateUser(int id, User user){
+    /*
+    Modificar usuario
+    true = correcto
+    false = no modificado
+     */
+
+    //falta arreglar si no se quieren cambiar todos los campos
+    public boolean updateUser(HttpHeaders header, int id, User user) throws UnauthorizedException, NotSuchPrivilegeException{
         Boolean a=false;
         String encriptedpass;
+        Verification v = new Verification();
         try {
-            if(user.getNombre()==null)
+            authManager.verify(header,v);
+            authManager.verifyCorrectUser(v, id);
+            if (user.getNombre() == null)
                 throw new NullPointerException();
             encriptedpass = td.encrypt(user.getContrasena());
             user.setId(id);
             user.setContrasena(encriptedpass);
             a = user.updatetUser();
+        }catch (UnauthorizedException ex) {
+            throw new UnauthorizedException("Unauthorized: user is not authorized");
+
+        }catch (NotSuchPrivilegeException ex){
+            throw new NotSuchPrivilegeException("Forbidden: User has not privileges");
+
         }catch(Exception e){
             logger.info("INFO: error al modificar usuario");
             a=false;
@@ -89,9 +112,21 @@ public class UserManager {
         return a;
     }
 
-    public User getUserById(int id){
+    public User getUserById(HttpHeaders header, int id)throws UnauthorizedException, NotSuchPrivilegeException{
+        Verification v = new Verification();
         User u= new User();
-        u.selectUserById(id);
+
+        try {
+            authManager.verify(header, v);
+            authManager.verifyCorrectUser(v, id);
+            u.selectUserById(id);
+
+        }catch (UnauthorizedException ex) {
+            throw new UnauthorizedException("Unauthorized: user is not authorized");
+
+        }catch (NotSuchPrivilegeException ex){
+            throw new NotSuchPrivilegeException("Forbidden: User has not privileges");
+        }
         return u;
     }
 
@@ -101,17 +136,38 @@ public class UserManager {
         return u;
     }
 
-    public User deleteUser(int id){
-        User u = new User();
-        u.selectUserById(id);
-        u.deleteUser();
 
-        return u;
+    public void deleteUser(HttpHeaders header, int id) throws UnauthorizedException, NotSuchPrivilegeException{
+        Verification v = new Verification();
+
+        try {
+            authManager.verify(header, v);
+            authManager.verifyAdmin(v);
+            User u = new User();
+            u.selectUserById(id);
+            u.deleteUser();
+        }catch (UnauthorizedException ex) {
+            throw new UnauthorizedException("Unauthorized: user is not authorized");
+
+        }catch (NotSuchPrivilegeException ex){
+            throw new NotSuchPrivilegeException("Forbidden: User has not privileges");
+        }
     }
 
-    public List listAllUsers(){
+    public List listAllUsers(HttpHeaders header) throws UnauthorizedException, NotSuchPrivilegeException{
+        Verification v = new Verification();
         List<User> list;
-        list = new User().findAllUsers();
+
+        try {
+            authManager.verify(header, v);
+            authManager.verifyAdmin(v);
+            list = new User().findAllUsers();
+        }catch (UnauthorizedException ex) {
+            throw new UnauthorizedException("Unauthorized: user is not authorized");
+
+        }catch (NotSuchPrivilegeException ex){
+            throw new NotSuchPrivilegeException("Forbidden: User has not privileges");
+        }
 
         return list;
     }
